requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
streamlit==1.38.0
pandas==2.2.2
numpy==2.0.2
scikit-learn==1.5.2
pyyaml==6.0.2
duckdb==1.1.3
requests==2.32.3
paho-mqtt==2.1.0
